buttons events moved to imageloader component and alt prop creation
