+++ src/stories/4-imageloeader.stories.tsx
@@ -1,5 +1,6 @@
 import React, {useState, useCallback} from 'react';
 import ImageLoader from '../components/ImageLoader';
+import { text } from '@storybook/addon-knobs';
 
 
 export default {
@@ -8,23 +9,15 @@ export default {
 
 export const base = () => {
 
+    const imagealt = text('Alt', 'Medicamento');
     const [imagesrc, setImagesrc] = useState("");
-
-    const loadImage = useCallback((event:any) => {
-        let file = event.target.files[0]; 
-        setImagesrc(URL.createObjectURL(file));
-    }, [imagesrc]);
-
-    const deleteImage = useCallback(() => {
-        setImagesrc("");
-    }, [imagesrc]);
 
     return(
         <div style={{width:"200px", height:"200px"}} >
             <ImageLoader
             imagesrc={imagesrc}
-            onLoadClick={loadImage}
-            onTrashClick={deleteImage}
+            imagealt={imagealt}
+            setImagesrc={setImagesrc}
             ></ImageLoader>
         </div>
     );
